fix: await async functions
