+++ README.md
@@ -51,7 +51,7 @@ If your device is among those affected and notifications are not displayed, plea
 
 ### Languages
 
-Florae is currently translated into the following languages: `English`, `Español` and `Français` thanks to Samuel Pernet.
+Florae is currently translated into the following languages: `English`, `Español`, `Français`, `Nederlands`, `中文`, `Русский` and `Arabic`.
 
 If you wish to contribute to Florae by adding a new language, just include the translation file in [`lib/l10n`](lib/l10n). I will be happy to accept your pull request.
