+++ lib/data/garden.dart
@@ -48,8 +48,7 @@ class Garden {
   Future<bool> deletePlant(Plant plant) async {
     List<Plant> allPlants = await getAllPlants();
 
-    var plantIndex =
-        allPlants.indexWhere((element) => element.name == plant.name);
+    var plantIndex = allPlants.indexWhere((element) => element.id == plant.id);
     if (plantIndex != -1) {
       allPlants.removeAt(plantIndex);
 
@@ -65,11 +64,9 @@ class Garden {
   Future<bool> updatePlant(Plant plant) async {
     List<Plant> allPlants = await getAllPlants();
 
-    var plantIndex =
-        allPlants.indexWhere((element) => element.name == plant.name);
+    var plantIndex = allPlants.indexWhere((element) => element.id == plant.id);
     if (plantIndex != -1) {
-      allPlants[allPlants.indexWhere((element) => element.name == plant.name)] =
-          plant;
+      allPlants[plantIndex] = plant;
 
       String jsonPlants = jsonEncode(allPlants);
       await store.setString("plants", jsonPlants);
